Remove one of dependency and introduce more interface for a more uniform experience
