+++ CoreOSC/CoreOSC.csproj
@@ -21,8 +21,5 @@
     <RepositoryType>git</RepositoryType>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="OneOf" Version="3.0.255" />
-  </ItemGroup>
 
 </Project>
