+++ CoreOSC/IOscPacket.cs
@@ -1,0 +1,6 @@
+﻿namespace LucHeart.CoreOSC;
+
+public interface IOscPacket
+{
+    public byte[] GetBytes();
+}
