+++ CoreOSC/IOscSender.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace LucHeart.CoreOSC;
+
+public interface IOscSender
+{
+    public Task SendAsync(byte[] message);
+    public Task SendAsync(IOscPacket packet);
+}
